Got project explorer working.
instead of returning the model indexes from the sub-models, we create persistent model indexes, store them in new indexes that belong to the MultipleProxyModel, and return the new indexes.

To prevent the Python garbage collector from deleting the persistent indexes from the sub-models, we store references to them in the MultipleProxyModel. Right now, this isn't done efficiently, so it needs to be changed.